# server/api/api-specification.yml
openapi: "3.0.0"
servers:
  - url: /api/v1
info:
  version: 1.0.0
paths:
  /me:
    get:
      summary: Returns the authenticated user
      operationId: UserInfo
      responses:
        "200":
          description: user response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "500":
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /poll:
    get:
      summary: Returns the authenticated user
      operationId: ListPolls
      responses:
        "200":
          description: user response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Poll"
        "500":
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

components:
  schemas:
    User:
      type: object
      required:
        - id
        - email
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string

    Poll:
      type: object
      required:
        - id
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        location:
          type: string
        options:
          type: array
          items:
            $ref: "#/components/schemas/PollOption"

    PollOption:
      type: object
      required:
        - id
        - start
        - end
      properties:
        id:
          type: string
        start:
          type: string
          format: date-time
          example: "2020-11-09T09:10:00Z"
        end:
          type: string
          format: date-time
          example: "2020-11-09T09:10:00Z"

    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string